Day15: study possible faster solution
The constant used by this problem are a bit too specific to be random.
It turns out that it could give optimization (at least for the reminder
by a mersenne number) but sadly in ruby the naive version is faster.
There is a huge performance hit by having a common implementation for
both part.

Change-Id: I3170b911b8629c8774b0e7db1510e9d6eccb5961